[ADD] Add module to edit the amount on the gift card and eWallet.
This module allows the user to input the amount to pay with gift card or eWallet. In the standard, if the gift card covers all the amount, the order is validated automatically. If the gift card does not cover all the amount, the total available amount will be used. This module lets you chose a lower amount to pay with the gift card.

[IMP] Added a condition to avoid the input popup to show when no lines are added.

[ADD] Change request for translation.

There was a change request to add spanish translations to the module in POS.
Also, on the pop up to input the amount the placeholder now is the maximum amount of the gift card.