[ADD] Add module to edit the amount on the gift card and eWallet.
This module allows the user to input the amount to pay with gift card or eWallet. In the standard, if the gift card covers all the amount, the order is validated automatically. If the gift card does not cover all the amount, the total available amount will be used. This module lets you chose a lower amount to pay with the gift card.